Minor OpenBlock/OpenSquare compile fix
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@112 37977431-3df6-0310-b722-df95706aa16b

+++ rts/PathEstimator.cpp
@@ -362,7 +362,7 @@ IPath::SearchResult CPathEstimator::GetPath(const MoveData& moveData, float3 sta
 		if(PATHDEBUG) {
 			*info << "PE: Search completed.\n";
 			*info << "Tested blocks: " << testedBlocks << "\n";
-			*info << "Open blocks: " << (openBlockBufferPointer - openBlockBuffer) << "\n";
+			*info << "Open blocks: " << (float)(openBlockBufferPointer - openBlockBuffer) << "\n";
 			*info << "Path length: " << (int)(path.path.size()) << "\n";
 			*info << "Path cost: " << path.pathCost << "\n";
 		}
@@ -370,7 +370,7 @@ IPath::SearchResult CPathEstimator::GetPath(const MoveData& moveData, float3 sta
 		if(PATHDEBUG) {
 			*info << "PE: Search failed!\n";
 			*info << "Tested blocks: " << testedBlocks << "\n";
-			*info << "Open blocks: " << (openBlockBufferPointer - openBlockBuffer) << "\n";
+			*info << "Open blocks: " << (float)(openBlockBufferPointer - openBlockBuffer) << "\n";
 		}
 	}
 	return result;
